chore: clean up host code
